feat: finish logout feature for user

+++ src/main/java/in/fssa/tharasworld/servlets/DeleteUserServlet.java
@@ -6,6 +6,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import in.fssa.tharasworld.entity.UserEntity;
 import in.fssa.tharasworld.exception.ServiceException;
@@ -29,15 +30,30 @@ public class DeleteUserServlet extends HttpServlet {
 		
 		try {
 			
-			UserService userService = new UserService();
+				UserService userService = new UserService();
+				String stringId = request.getParameter("id");
+				HttpSession session = request.getSession(false);
+				if (session != null) {
+				    session.setAttribute("userId", 0);
+				    Object userIdAttribute = session.getAttribute("userId");
+				    System.out.println("userIdAttribute: " + userIdAttribute);
+				    session.invalidate();
+				}
 			
-			String idParams = request.getParameter("id");
-			
-			int id = Integer.parseInt(idParams);
-			
-			userService.delete(id); 
-			
-			response.sendRedirect(request.getContextPath() + "/user_records");
+				if (stringId != null && !stringId.isEmpty()) {
+					int id = Integer.parseInt(stringId);
+					userService.delete(id);
+					response.sendRedirect(request.getContextPath() + "/index.jsp");
+				}
+
+				
+//			String idParams = request.getParameter("id");
+//			
+//			int id = Integer.parseInt(idParams);
+//			
+//			userService.delete(id); 
+//			
+//			response.sendRedirect(request.getContextPath() + "/user_records");
 			
 		} catch (ValidationException e) {
 			e.printStackTrace();
